Add `sparse_set` to `git_repository()`.
This implies a partial fetch of the repo, downloading only the files necessary.
This will require support from both the local git client and the server. It can
be slightly slower for a small repository, but orders of magnitude faster and
more space-efficient when you need only a small slice of a very large
repository.