Add `fallback_to_full` attribute
This makes it possible to disable falling back to a full fetch in the event
that a shallow or partial fetch fails. This could happen due to lack of support
from either the remote or the local Git (sometimes due to an old version).
However, for very large repositories, the consequences of a full fetch are
significant, so it may be preferable to error out early and allow the user to
address the issue (e.g., update their local Git).